html/dom/area-element-usvstring.html

<!DOCTYPE html>
<html>
  <head>
    <title>area ping test for USVString</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    test(function () {
        var a = document.createElement("area");
        a.ping = '\uD999';
        assert_equals(a.ping, '\ufffd');
    }, "unpaired surrogate codepoint should be replaced with U+FFFD");
    </script>
  </body>
</html>
